Fixed placing and destroying blocks
The previous code had a bug where the coordinates for placing and destroying blocks was wrong. This is now fixed. The chunks are currenlty re-generated each time, so if blocks are placed or destroyed, then the player goes away and comes back, the modifications will be gone. Next thing to implement is to mark the modified chunks in memory and store these in a cached chunk array that will give the mesh of the chunk instead of regenerating it.

+++ include/game_config.h
@@ -20,7 +20,7 @@ typedef struct {
     float sensitivity;
     float speed;
     float fov;
-    float render_distance;
+    uint8_t render_distance;
 
     /* World settings */
     float max_reach;
